requirements.txt
pytesseract
pymupdf
pillow
reportlab
pandas